extended display of GSE, added search and button to re-import
The display of the grade span expectations includes more related
data: domains, unifying themes.

The index page has a search method (only looks at grade-span for now).

A button to re parse the gse xml doc is available now also. This should
be limited to an admin role in the controller and view. The role_requirements
plugin doesn't appear to work in the view. I'm looking at switching to
simple_access_control (which we use in the DIY and SDS).