Use will_paginates Collection#total_entries attribute
Instead of always displaying the total number of GSEs, display
the total in the Collection object returned by will_paginate. This
is the same number if no other selection criteria are used to limit
the result set, however if a search string is used that returns a
subset of the records Collection#total_entries will return the size
of that collection.