update user model for new restful_auth, fix failed login
The validate class methods of restful_authentication were updated and renamed,
the user model is updated to reflect this.

I also have started the removal of open-id in this section of the user model.

Also, if a login failed the current_user wasn't being set to the anonymous
user which was causing an error in the view