refactor page view system, sort action in controller
The edit template in the page view system now calls specific
page element partials directly instead of using html_for_element
method in PageHelper. This looks like this:

  <%= render(:partial => element.embeddable_type.underscore,
    :object => element.embeddable) %>

InvestigationStepsHelper is now PagesHelper but so far it's
has not been needed.

I renamed the edit partials to remove the 'edit_' suffix. This way
the object being passed in can be referred to with the name of
the partial. So the new partial name _xhtml.erb starts like this:

  <% form_for(xhtml) do |f| %>

I added tinymce in application_html.haml but it is not being applied
to the textareas visible when editing a page.

Also added a route for sort_element to the page resource and refactored
the controller action -- this has not been tested.

Changed text: "sortable_step" to "sortable_element" generally.

Most of the actions in the page_controller now start like this:

  @page = Page.find(params[:id], :include => :page_elements)
  @page_elements = @page.page_elements

The :include loads the page_elements in the same SQL request
making this much more efficient.

I re-enabled textmate-footnotes ... if this is still a problem for
you we can make the loading conditional on an additional yaml config
that is not checked in.