more refactoring of the pages_controller show action and views
The xhtmls and open_responses controllers now just return
partials for their respective embeddables for their new actions.

The two new partials use haml.

This is linked up to the new xhtml and open response links on
the show action for a page.

Alternate forms of the routes for the embeddables: xhtmls, and
open_responses were created that support specifying the page_id:

Here's an example:

  map.resources :xhtmls, :path_prefix => '/pages/:page_id', :name_prefix => 'page_'

This results in routes like:

  /pages/:id/xhtmls/new

Which can be specified in the Ruby code like this:

  new_page_xhtml_path(@page)

The link_to_remote links on the show page action have been
extended to include the element to update, the position in the container
to insert the new html and a javascript Scriptaculous call to re-create
the Sortable container. Without the last step the new items inserted
are not sortable.

The create actions on the xhtmls and open_responses controllers
now include a conditional creation of the association between
an embeddable and it's parent page. This function is performed
when a page_id attribute is present in the params hash. This work
has not been tested yet.

Moved the javascript initialization back to the beginning of the
page because having it at the end was causing FireFox to get into
a loop and bog down -- problem didn't occur in Safari ???