refactored controllers, views, and partials for embeddable components
This work covers: data_collectors, xhtmls, and open response
embeddables.

The page show view uses a new: render_partial_for(embeddable)
helper which now directly calls the <embeddable_name> partial
in the view folder for the embeddable.

Examples:

  data_collectors/_data_collector.html.haml
  open_responses/_open_response.html.haml
  xhtmls/xhtml.html.haml

These partials display the menu and content divs that go inside
an item div.

In the menu div these partials display the name, a
button (using button_remote_for) labeled "Edit", and the
content.

Clicking the "Edit" button makes a request to the remote_form
partial for the embeddable which generates the replacement
inner_html content for the item container.

The menu on the form partial includes buttons named "Save" and
"Cancel".

Some problems:

Previously the partials were including the item container in
their rendering -- this meant that after an edit cycle the container
was shifted to the right because now one item container was nested
inside the original.

After I fixed this the item div is not updated when clicking Edit.

I implemented the data_collectors controller so this same sequence
can happen directly with a Data Collector instance. This still
works but it isn't obvious what the difference is.

Another problem. When editing an xhtml I haven't been able to get
TinyMCE to appear. I've added the creation of a new tinymce
instance to the "complete callback to no obvious effect.