update gem configs in environment.rb
I needed these changes to get the app starting
in jruby.

hpricot v0.7 does not yet work with jruby so
spec 0.6.164

grit has a dependency on open4 but for whatever 
reason in jruby open4 is not installed when
installing mojombo-grit

haml should be 2.0.7 or greater (perhaps it should
only be greater -- the most recent is 2.0.9).

The spec for prawn-form was incomplete and 
commented-out -- it should work now.