refactor the views for xhtmls
In order to get the views working correctly I 
worked on just the xhtml views. There is a new
show partial which is called by the show view AND
rendered by the controller update action in response 
to an XHR request.

The show partial can probably be generalized into 
a shared partial ...