sections tabs are shown when pages are being edited
section tabs now support more tabs that need
multiple lines to be displayed.

working towards getting the section meta-data
displaying in a summary/expanded and edit forms

the 'active' class is now used on the containing
li elements (it has not been deleted yet from
the a elements but it probably could be).