scoped model access to ITSI activities, models, model_types
This now allows read access to ITSI resources.

The purpose is to support importing ITSI activities.

A config like this should be added to database.yml
pointing to the host and database for ITSI.

itsi: 
  encoding: utf8
  adapter: mysql
  host: localhost
  username: root
  database: production_itsi_prod
  password: ******
  pool: 5

Then itsi activities can be referenced like this:

a = Itsi::Activity.find(:first)

a.name
# => "MW Sandbox: Demo MW capacity"

a.user.name
# => "itest itest"

p = a.probe_type
# => #<ProbeType id: 8, user_id: nil, name: "Relative Humidity", ptype: 7, 
#     step_size: 0.1, display_precision: -1, port: 0, unit: "percentage", 
#     min: 10.0, max: 90.0, period: 0.1, uuid: nil>

p.activities.length
# => 16

a.probe_type.name
# => "Relative Humidity"

The instances in the ProbeType and VendorInterface tables in
RITES have the same primary key ids and equivalent functions
as the tables in ITSI so the ITSI models use the RITES tables
for these associations.