styling updates to investigations, sections, pages
I'm using item_collection to style collections
of investigations, sections, pages to make these
lists lighter weight.

by default descriptions/preview detail divs
open when load a section or page

action_menu inside an item_collection is not as wide.

I added a how/hide toggle all link to the top of 
collections.

I also added an edit link at the top which duplicates 
the function of the link in