added print action to investigations and sections
Accessed like this:

  http://localhost:3000/investigations/6/print

There are no links rendered into the view yet so
you need to construct the url manually.

Much work is still needed.

When printing to a pdf on a Mac the Canvas graphs 
render but do not appear in the correct position.

Also the techniques described here (and presumably
newer articles on the same subject) should be 
applied.

  http://www.alistapart.com/articles/goingtoprint