page position scoped to section, Page#name method added
I added the instance method Page#name to return the string
'untitled page' if the name was nil or empty and
discovered that the position atttribute for new pages
is not being set correctly.

I updated the page model acts_as_list to be scoped to 
a section but this didn't fix this problem yet:

position attributes set in correctly when pages are 
made in a new section
http://jira.concord.org/browse/RITES-82

I tried adding @page.save to the add_page action in 
the sections controller butt hat didn't help either.