RUN MIGRATION!!! change attribute j2se to j2se_version
The j2se attribute on the real Jnlp object is
the hpricot document created when parsing.

This attribute in the rails model should instead be
j2se_version -- which is the value of the version
attribute for the j2se element in the original jnlp.

If you have already created jnlps you should delete
and recreate them.