RUN MIGRATIONS!!! add otrunk_examples introspection
Creates a set of models, libraries, and tasks that support
introspection into the imports and view_entry patterns
actually used in the otml files in otrunk-examples.

New Rake task:

rake rigse:import:generate_otrunk_examples_rails_models

An OtrunkExample::OtrunkImport instance has these 
new instance methods:

  all_associated_imports

Returns the union of all the imports in all the 
otml files that also use this import.

  all_associated_view_entries

Returns the union of all the view_entries in all the 
otml files that also use this import.

  standard_view_entry
  standard_edit_view_entry

Returns the standard and edit views for the import.
This is done using a simple heuristic that adds either:
'View' or 'Edit View' to the end of the import name.