add task: delete_users_and_restore_default_users_roles
This can be used if the users table gets crunched 
somehow while testing. It uses an SQL truncate
command on the users table so when the default users
are re-created they should have the same primary keys.

This is more useful to a developer if they have 
previously set the admin user in config/settings.yml.

If this is the standard user they login as and the
setup scripts were run previously to create this 
admin user then the user will once again have a 
primary key value of 1.