placeholders for creating has_many :investigations
These are placeholders for exposing the
relationship through a singular AR association.

The finder_sql will need to be completely changed.

Once this is working it should be moved to a library
which can be used to extend all the embeddable models.

Perhaps an abstract class they all inherit from,