this prevents a conflict if uuidtools 2 is installed
When rails starts up it first loads it's rake tasks
and then loads the application rake tasks.

When the main app rake task executed:

  require 'uuidtools'

RubyGems loaded to newest version of uuidtools
found in the active local gem repositories (which
in the case of the production instance was v2.0.0 
because that version was required for the staging
instance which runs on the same server).

By executing:

  gem "uuidtools", '= 1.0.7'

before the require statement this instance of the 
production instance of the RITES codebase will load 
v1.0.7 of uuidtools instead.

Then later in the startup process in the 
Rails::Initializer.run do |config| block in
config/environment.rb when the the gem dependency
for uuidtools is stated:

  config.gem "uuidtools", :version => '= 1.0.7'

no conflict will occur.