turn on editing embeddables in Java for everybody
If an embeddable responds to: 

  embeddable.authorable_in_java?

with a value that is true (anything not nil or false)
then the Page authoring UI will display both a
regular pencil icon and the Pencil Icon with a small
'J' for editing in Java.

This is only operational for editing the Graphs 
at this point for regular users.

However it is also operational in the raw_otmls 
controller which is not exposed in the UI.