task default_users_roles creates portal resources
The teacher and student roles are no longer used
and will be deleted when this task is run.

An author user is created and one default Investigation
is created for that author. The publication status
for this Investigation is set to 'published'

Teacher and student users are created along with 
the associated portal_teacher and portal_student
resources.

A default district and school are created using a 
name from new parameters in config/settings.yml --
an updated sample is in config/settings.sample.yml.

Two semesters and one grade_level (9th
grade) are created for the default school.

One default course and class are created for the 
school and teacher.

An offering is created for this class referencing
the default Investigation created for the author user.

The portal_student is assigned to this class.

The creation of all default users can now be done
with with one key response instead of confirming each
attribute.