fix setting of user roles in controller and views
I've backported from the with_portal branch these
two methods for a User object.

  User#add_role(role)
  User#remove_role(role)

In addition I've added this new method:

  User#set_role_ids(role_ids)

role_ids is an array of role ids as integers or
strings

The method will enable the roles in the role_id list
and disable the roles not in the list -- without
making duplicates.