task default_users_roles creates SdsUsers and Learner
The rake task:

  rake rigse:setup:default_users_roles

Now creates SdsUsers for all the standard default 
users (if they don't already exist).

This functionality is dependent on new code in the User
model in the portal engine.

In addition a learner instance is created for the 
default_student in the offering for the default_class.