updated db/schema.rb, removed nces tables again
The nces elements in the schema still keep getting
checked in -- that's because after the code was 
moved to the portal engine I ended up just deleting the 
code and migrations.

So folks that have the tables:

  nces06_districts
  nces06_schools
  
should delete them by hand from your database.

The nces code and mirgations were removed in:

  e449415bd285b2779d627de5e2878a4573f2e615