additions to settings.yml for maven-jnlp artifacts
Multiple maven_jnlp_servers can be defined in:

  config/settings.yml

in addition there are values for:

  default_maven_jnlp_server
  default_maven_jnlp_family

If you setup a local jnlp server you can now reference
this server. In addition you can specify a different
maven_jnlp_family. The default_maven_jnlp_family is:

  all-otrunk-snapshot

I am using the following to specify two maven_jnlp_servers:

  default_maven_jnlp_server: local  default_maven_jnlp_family: all-otrunk-snapshot  maven_jnlp_servers:    -       :name: concord      :host: http://jnlp.concord.org      :path: /dev/org/concord/maven-jnlp/    -       :name: local      :host: http://jnlp      :path: /org/concord/maven-jnlp/

If you add a new jnlp_server you will need to rerun
this rake task:

  rake rigse:jnlp:generate_maven_jnlp_family_of_resources

The task is not yet smart enough to know if you have
already generated model instances for a jnlp_server
so if you are doing this first delete the existing
maven_jnlp resources with this task:

  rake rigse:jnlp:delete_maven_jnlp_resources