reference OTDataGraphable instead of OTDataStore
By referencing OTDataGraphable when combing data
from another graph we preserve the color of the original
data. So in the common situation of an original
prediction graph (which defaults to red) which is
later combined with a sensor graph (which defaults
to blue) the overlay from the prediction graph now
will be red.