TruncatableXhtml: Xhtml, OpenResponse, and MultipleChoice
Generalize the pattern of extracting and sof-truncating
text from xhtml content to use to generate the name 
attribute.

Xhtml models process the content attribute, while 
OpenResponse and MultipleChoice models process the 
prompt attribute.

The Module is now named TruncatableXhtml and it adds
two methods to models it is Included in:

  truncate_from_xhtml(attribute, limit=24, soft_limit=8)

Extracts and soft truncates text from an xhtml document stored
in the an attribute for models this module is included in.

  soft_truncate(string, length, soft_limit=nil)

Truncates a string to length characters with an additional
optional soft limit. If a soft_limit length is specified  
and a word in the string crosses the length boundary the
length of the returned string will be increased up to a
total length of (length + soft_limit) in an attempt to
complete the last word.