add OTHoltsMethodSmoothing to data_filters.yaml
Smooths the data. 

Default values for coefficients: 

  k0 (level smoothing): 0.5
  k1 (trend smoothing): 0.5. 

Run rake task:

  rake db:backup:load_probe_configurations

or cap task for remote update:

  cap import:reload_probe_configurations

This filter is use in some of the LOOPS activities to
smooth the data used for showing velocity graphs.

Example:

LOOPS/fm-4-activity.otml has an OTDataCollector that
uses this nested filter structure:

dataProducer
  OTHoltsMethodSmoothing
    OTTimeLimitDataProducerFilter
      OTDifferentiatingProducerFilter
        OTLinearProducerFilter
          OTSensorDataProxy