Add admin/project controller, model, views
So far a Project is an object that keeps track
of the maven_jnlp server, family, jnlp_version,
and if snapshots are enabled.

Soon a Project will relate to more object collections
such as Investigations and Districts.

The JnlpAdaptor initialize method now requires a
project object.

There is a class method to create the first 
(and default) project:

  Admin::Project.create_or_update_from_settings_yml

This requires that a valid config/settings.yml exists.