page caching for investigation otmls
The InvestigationSweeper is not following the pattern
described in the rails documentation. When it is called 
it has not access to the controller and the various
expire_xxx methods are nil.

So the path to the otml to expire is hard-coded.

Since it has no access to he controller when it is 
called it could just as well be an model observer.