support running jnlps in preview mode all in RITES
This commit replaces the use of the SAIL Data Service
for running Jnlps in preview mode.

There is a new Mime::Type named :config

The runnable controller show actions now have a line
like this in their respond_to block:

format.config { render :partial => 'shared/show', :locals => { :runnable => @investigation } }

There are new shared partials for both *.jnlp and 
*.config show requests:

  shared/_show.jnlp.builder
  shared/_show.config.bulder

Also replaced the use of the name: :runnable_object
with :runnable for consistency in all the runnable
controllers and in the element generator.