cap deploy now uses the dataservice_rails_2_3_3_portal branch
The dataservice_rails_2_3_3_portal branch is now 
deployed to the rites dev server instead of master.

This branch folds the portal engine back into the
rites code and implements a dataservice that replaces 
the SDS.

Uploading learner data does not work yet -- the rack
middleware I started lib/expand_b64_gzip.rb needs 
to be finished first.

The Java client first gzips and then encodes that binary
data with b65 to send in a post. Right now rails 
chokes because it also assumes the upload is XML 
(it will be after it is decompressed) and tries
to run Hash.from_xml on the b64gzip data.

The filter should just look to see of the input
is of type b64gzip and if so decompress it and leave
the result for later in the filter chain.