Add SearchableModel and class_method: display_name to Portal models
The display_name class method is needed for integration
with the RITES component editing system.

Portal models are only extended with SearchableModel
if they have actual searchable attributes that aren't 
delegated away.

When we have a better searching system we'll be able
to search models like Portal::Learners which only
have associated attributes.