extract otml from bundle_content when created
The otml in an uploaded bundle is extracted when
the bundle_content model is created.

If the bundle_content model has not been processed
this function is also performed when the model is 
saved.

Once processing has taken place the ivar 'processed' 
is set to true.

Right now the content is in effect duplicated in both
the body attribute, which holds the entire session
bundle and the otml attribute which holds the otml
expanded from the value in the first sockEntries 
element.

The following fragment of the extract_otml method:

  body.slice(/<sockEntries value="(.*?)"/, 1)

was originally:

  body.slice!(/<sockEntries value="(.*?)"/, 1)

Which removed the b64gzip string from the value
attribute in the body.

However until I can supply a Java OTrunk instance
with otml instead of a bundle this just made the
reconstruction of the eportfolio session bundle more
difficult.

So for now the table is twice as big as it should be.

In addition some of the method names mixed in with 
the SailBundleContent module have changed.