git_repo_info helper deals with non-HEAD checkout
This helper which returns an array of items to help
rendering info about the current commit the app is 
using was broken for a while.

This turned out to be more trouble than a normal bug
when using trying to do a git bisect to uncover 
the introduction of a different bug.

During the bisect the app is not actually on a HEAD
branch and this helper was throwing an error ... which
meant that testing to determine if the new error was
present became much harder.

I ended up creating a branch over and over and did the bisect by hand.

  git co master
  git branch -D test
  git co -b test <sha1 hash>
  script/server -p 3000

  ... test in app

This is obviously something that should be avoided.

What this means is that it is VERY important to not checkin bugs 
in this helper because that then ends up making finding other 
bugs much harder.