MIGRATION: add portal/grade model et al
There are a fixed number of Portal:Grade objects
representing the grades kindergarten through 12th grade.

These objects are setup by the rake task:

  rigse:setup:default_users_roles_and_portal_resources

There is a new value in config/settings.yml that 
allows the active grades for the site to be listed:

The default setting for RITES are grades 6-12:

  APP_CONFIG[:active_grades]
  => [6, 7, 8, 9, 10, 11, 12]

The 'active' state of the Portal::Grade objects are set 
to these values when the app is started.