render HttpCookieServiceImpl service in show config
This new service will allow the config document to pass
http session info to the Java OTrunk instance so that the responses
to subsequent requests from the Java OTrunk instance to the web 
application can be customized for the user.

See: https://confluence.concord.org/display/CSP/Design+for+passing+session+info+to+SAIL

The use cases include

* customizing the Vendor Interface selection for the user.
* protecting privacy of data.

To be robust a method for adding a one-time use token to
the url used to generate the config response will also be needed.