simplify gem installation when creating RITES app
Fewer gems need to be installed before config/setup.rb
can be run to completion.  The gems in vendor/gems
are added to the load path when running setup.rb so
uuidtools and highline can be loaded from the app instead 
of being installed into the Ruby system gem location.

In addition a few system gem dependencies are added.
These are differentiated into Ruby and JRuby dependencies.

Lastly after completion of setup.rb the user is reminded
to run the rails rake command:

  rake gems:install

Which will install gems which require native compilation.

Before continuing the install with:

  rake rigse:setup:new_rites_app