IMPORTANT: new elements for settings.yml
Update your config/settings/yml from teh changes in this sample.

Processing and importing of the otrunk-examples can be disabled:

  include_otrunk_examples: false

I'm not sure whether this is a problem for RITES. Disabling 
it makes the creation of an initial app faster.

Active Grades must now be defined as strings:

  active_grades:
    - "6"
    - "7"

Active School Levels: only the school levels defined here
will be imported by the task:

  portal:setup:create_districts_and_schools_from_nces_data

  1 = Elementary
  2 = Middle School
  3 = High School
  4 = Undefined

It is possible that undefined could be used in the NCES data
to represent schools that combined two or more of the first
three categories.

  Example:

  active_school_levels:
    - "2"
    - "3"
    - "4"