add grade-matching methods to Nces06School
  Nces06School#active_grades

Returns a list of active grades for the NCES school in the same
format as that produced by: APP_CONFIG[:active_grades]

  Nces06School#grade_match

Is an attempt to create a boolean method comparing
an array of grade strings to the active_grades for an
NCES school.

It is not yet robust enough however, consider:

  APP_CONFIG[:active_grades] => ["6", "7", "8"]
  Nces06School#active_grades => ["4", "5", "6"]

  Nces06School#grade_match => true

The current system uses APP_CONFIG[:active_levels] instead
to filter which Schools are created from the NCES data.