SPLIT UP TASK: default_users_roles_portal_resources
The task:

  rake rigse:setup:default_users_roles_portal_resources

Is now split into two separate tasks:

  rake rigse:setup:default_users_roles

  rake rigse:setup:default_portal_resources

This was done so that when running the task:

  rake rigse:setup:new_rites_app

The task: 

  rake db:backup:load_probe_configurations

could be invoked between the two new tasks.

Loading the probe configurations requires the administrative user
to be created.

Creating the default investigation requires the probe 
configurations to be loaded.

These problems only occur when building from scratch.