NEW config file: config/rinet_data_config.yml NEW class: RinetData
Create a file: config/rinet_data_config.yml
from the template in config/rinet_data_config.sample.yml

The configuration data specify where to get the RINET
csv data for Districts and Schools.

The district list is in a class variable: @@districts

  rd = RinetData.new
  rd.get_csv_files
  pd = rd.parse_csv_files

The result is an array of the data parsed from the csv files.

Still to be completed is processing the imported data
and creating the associated RITES resources.

Using gems:
  net-sftp: http://net-ssh.rubyforge.org/sftp/v2/api/index.html
  fastercsv: http://fastercsv.rubyforge.org/