don't run rites.rb initializer if testing or rake task
The initialization of the default Admin::Project model with
data from settings.yml is probably only necessary when starting
the RAILS environment either as a server or with script/runner.

Now skip this initialization if:

  RAILS_ENV == 'test' || $PROGRAM_NAME =~ /rake/

If this strategy works the rest of the code here that works
to avoid throwing errors when the database either doesn't exist
or is empty can probably be removed.

Discovered presence of $PROGRAM_NAME by dropping into Ruby debugger
in rites.rb just before initialization code executed and typing
the debugger command:

  var global

many interesting bits there I didn't know about.