README: regularize <pre><code>...</code></pre>
The textile formatter for code:

  bc. 

works well ... it just wraps the following block content
in: <pre><code> tags.

A limitation is that the first blank line ends
this kind of block.

A work-around is to include one or more spaces in blank lines
that should NOT end the code block.

A second gotcha is that in a situation like the following:

bc. 
  # A code comment followed by a blank line
  
  # and another code comment

There must be at least one space character after: 'bc.'