update to latest highline 1.5.2 trunk
see: http://github.com/stepheneb/highline

If unix try: termios, ffi-ncurses, then stty

If running highline in unix first try loading termios.
If that fails try ffi-ncurses (this is the only workable
option right now for JRuby), if that fails use stty.

Use Ncurses.cbreak to support single char input

Without this modification when ffi-ncurses is used in JRuby
only the first of the following two agree invocations
returns after a single char input. The second invocation
waits for a newline.

  agree("Are you sure you want to do this?  (y/n) ", true)
  agree("Are you really sure you want to do this?  (y/n) ", true)

Use stty when ffi-ncurses not present in JRuby

If ffi-ncurses is not available AND highline is running in
JRuby fall back to using stty and print a suggestion to
install ffi-ncurses to STDERR.

There are character input issues but this allows successful
loading of gems like net-ssh which have a highline dependency.