fix users fixtures, 'mock, REST_AUTH_SITE_KEY
In order to run the user specs the encrypted passwords
for the 'quentin' and 'aaron' users in spec/fixtures/users.yml
need to be created with a REST_AUTH_SITE_KEY used for testing.

spec/support/mock_models now overwrites any existing
REST_AUTH_SITE_KEY with one for testing.