user create: separate test for login from email
If the row has a [:login] value try to find or create
the user. NOT only if the row has both [:login]  and 
[:email].

Moved some higher level methods just under the initialize
method. These will need refactoring to deal as clearly
as possible with the different ways to indicate the
districts that should be processed.

  RinetData.new()

uses parameters in rinet_data.yml

  RinetData#run_importer(district_dir)

Uses the district passed in.