add NCES-related class methods: Portal::District, School
Given an NCES local school id that matches the SEASCH field in an NCES school
find and return the first district that is associated with the NCES or nil.

example: 

  Portal::School.find_by_state_and_nces_local_id('RI', 39123).name
  => "Woonsocket High School"


Given a school name that matches the SEASCH field in an NCES school find
and return the first school that is associated with the NCES school or nil.

example: 

  Portal::School.find_by_state_and_school_name('RI', "Woonsocket High School").nces_local_id
  => "39123"

if the school is a 'real' school return the NCES local school id

  school.nces_local_id
  => "39123"

Given an NCES local district id that matches the STID field in an NCES district
find and return the first district that is associated with the NCES district or nil.

example: 

  Portal::District.find_by_state_and_nces_local_id('RI', 39).name
  => "Woonsocket"

Given a district name that matches the NAME field in an NCES district find
and return the first district that is associated with the NCES district or nil.

example: 

  Portal::District.find_by_state_and_district_name('RI', "Woonsocket").nces_local_id
  => "39"

If the district is a 'real' district return the NCES local district id

  district.nces_local_id
  => "39"