prevent deletion of investigation in use by clazzs
As reported by Ed Hazzard, if an offering exists 
for a clazz, and the investigation is destroyed
there was an NPE.

So now we Check if the investigation is in use by
a clazz before destroying it.

tests should be made for a variety of scenarios
like this.

Changed files:

  - investigation_controller checks for offerings.
	
  - destroy.js.rjs shows flash message in rjs

  - application.html.haml fixes flash displaying.