update the gem specifications for vendored gems
These were produced after updating the jnlp gem to v0.0.5.3
by running the following task in MRI:

  rake gems:refresh_specs

It's not clear why all the specs were are different than the 
previous ones in the repository.

It may be that both MRI and JRuby produce lexically different 
but semantically identical gem specifications.