gui-tester jnlp now part of generated jnlp resources
The gui-tester jnlp family is now *always* included when generating
maven_jnlp resources for an Investigations instance.

To update a running instance use this new rake task:

  rake rigse:jnlp:delete_and_regenerate_maven_jnlp_resources

To run this task *without* confirmation in the console prefix
the invocation with:

  ANSWER_YES=true 

The task jnlp.rake now uses this method: wrapped_agree
to query for console confirmation.  If the environmental
variable: ANSWER_YES is set then the console confirmation
is skipped and the answer is presumed to be 'y'.

There are some new class methods for deleting cached MavenJnlp 
resources which are referenced in this class method:

  MavenJnlp::MavenJnlpServer.delete_all_cached_resources

This method deletes the yaml caches for the mMavenJnlp and Jnlp objects
saved in:

  config/maven_jnlp_object_<n>.yaml
  config/jnlp_objects/*