fixes to projects controller
In working on getting the rspec controller tests to work
it was complaining about the render_scope method called in a
before_filter. I can't see that needed in this controller,
or how it worked at all ??? 

It doesn't appear to be in scope here.

In addition the rspec tests complained about:

  redirect_to(projects_url)

Which was a bug -- since this is Admin::Projects it needed to be:

  redirect_to(admin_projects_url)