check for existence and length > 0 of config files
Whne running config/setup.rb on a new app dir structure setup
by capistrano files like config/settings.yml exist but they
are sym-links pointing to zero length files in shared/config.

The issue isn't the sym-link it's the fact that the setup
program was assuming that if the file existed it had content.