restore initial display of learner responses
Restore portion of offering_for_teacher view partial
that displayed learner response to open-response questions.

As part of the prep for the merge of sparks back into master
I had copied the offering_for_teacher partial into the 
sparks theme and restored the default copy because some
elements of the partial were commented-out because they 
were hard-coded for a top-level-container of Investigations.

This commit restores just the portion that displayed 
learner data.