more ways to refresh learner.saveable_response_objects
In clazz controller show and add_offering actions, also
in offering and class models.

FIXME: this needs to be made faster by only doing this kind
of update when changes actually occur.

Changes in the list of all saveable objects for a learner
occur when 

1. A new embeddable object that has a corresponding saveable 
is created in the runnable the learner uses.

2. An existing embeddable object that has a corresponding saveable
is created in the runnable the learner uses.

3. An existing embeddable object that has a corresponding saveable
is moved in the runnable the learner uses. Saveables are (or should be)
ordered by corresponding position in the runnable.

4. A learner is added to an existing offering.

5. A new offering is created for a class with more than 0 students.

FIXME: When an authopr wants to delete and authorable object 
that has corresponding learner data the object should be archived 
instead of deleted so the context for the saved learner data can 
be re-created.