MIGRATION: saveable answers update counter_cache
The saveable root objects: OpenResponse and MultipleChoice
now have a counter_cache attribute named :response_count
that is incremented when answers are added. This makes 
finding the number of times a learner has responded
to a unique question faster.

In order for ActiveRecord to manage the updating of the
counter_cache automatically new answers should be created
through the association.

An example of adding a new answer to an existing 
Saveable::OpenResponse object through the association:

  open_response.answers.create(:bundle_content_id = 3912, :answer => 'yes')

The counter_cache in the open_response object will be 
automatically incremented by ActiveRecord.

I have not applied this pattern to the 
Saveable::Sparks::MeasuringResistance object, there may 
be a few more details to be considered in that case that 
I am not considering just yet.