Reports now use Report::Util object to ask for learner saveables, etc.
A Report::Util instance should be instantiated by calling Report::Util.factory(offering).
It will initialize and return a Report::Util instance that has the offering's data loaded.
If a Report::Util instance has already been created for that offering, it will return the previously created instance, making calls to the factory class method cheap and avoiding the need to pass around an instance of the Report::Util object.