controller integration tests for Embeddable::DataCollector
I'm using rspec-rails integrate_views method so that these
controller tests also run the code to render views.

This is a different way of doing what approaches integration 
testing -- however it is completed in the controller specs.

I'm testing the structure of all the alternative (non-html) 
types of formatted views produced by calling the controller.

  GET otml
  GET dynamic_otml
  GET config
  GET jnlp

and:

  GET :edit otml
  GET :edit dynamic_otml
  GET :edit config
  GET :edit jnlp

This now correctly flags: GET :edit otml as failing.

Actually there are a bunch of failures but all except the
one I just mentioned will be fixed by the next commit.