refactor ApplicationHelper spec
The old form worked but caused bad side effects by
mixing the methods in the ApplicationHelper module
into Object globally. This caused other tests to fail
when the code they were testing expected to be able to
make rational choices by reflecting on whether an object
was able to respond to a method.