Features use steps to create default project/jnlps
You can use now use steps for generating the default project
and jnlp resources with either mocks or factories

You can start a feature with either of these steps:

  Given The default project and jnlp resources exist using mocks

  Given The default project and jnlp resources exist using factories

I haven't tried using both at once.

Currently the only feature that needs factories is:

  admin_configures_vendor_interfaces

This is because the external_user_interface method is not mocked.