move investigation duplication to separate spec file
The duplications step duplicates the investigation
in a before(:all) block instead of in a before(:each) for 
each of the 11 tests.

This is about 4 times faster but sharing state between
tests can introduce hard-to-find bugs in the test suite.

So I moved the duplication tests into their own spec file
to isolate the before(:all) block from other unit tests 
for the Investigation model.